requirements.txt
google-genai>=1.0.0
pinecone>=5.0.0
pymupdf>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-adk>=0.1.0
